move custom reports as a "common" feature
